requirements.txt
boto3==1.17.18
botocore==1.20.18
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
ipaddress==1.0.23
ipdata==3.4.3
jmespath==0.10.0
pdfrw==0.4
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
tqdm==4.58.0
urllib3==1.26.3
